add RuleDao.selectById to be used by Views plugin

+++ sonar-db/src/main/java/org/sonar/db/rule/RuleMapper.java
@@ -38,7 +38,7 @@ public interface RuleMapper {
 
   List<RuleDto> selectBySubCharacteristicId(int characteristicId);
 
-  RuleDto selectById(Integer id);
+  RuleDto selectById(long id);
 
   RuleDto selectByKey(RuleKey ruleKey);
